Added an IndexedMzML class to make random access of indexed MzML documents possible.
New class is derived from the original mzml.MzML reader, with a modified __init__ method
to build up a byte offset index for spectra relative to the start of the file. If applied
to a non-indexed MzML file, the class will fall back on its parent class methods with no
loss of functionality, but without the substantial improvement in random access performance
by `get_by_id`. As a convenience notation, I also set the __getitem__ method on the INdexedMzML
class to invoke `get_by_id` as the interface makes sense given the behavior, but this asymmetry
with the base class may not be worth the saved characters.