Redefined a generic line tokenizer to construct the byte offsets for any type of XML file, regardless of white space.
Added an IndexedXML base class for later byte offset indexed XML file readers
This type supplies no new public methods, but redefines internally how :meth:`get_by_id`
functions, using a fast look up when the id is in the offset index, and falling back to the
normal behavior when an index is not available.

Added an IndexedMzIdentML class inheriting from MzIdentML and IndexedXML. This class builds an
Index for a subset of the elements used by reference in the parsing process to speed up parsing.

In order to justify the process, I compared the performance of the IndexedMzIdentML parser with the
base MzIdentML parser with in-memory caching of lxml.Elements turned on, and speeds were comparable,
but with a 33% reduction in memory consumption. Tuning of the elements which are indexed could produce
is replaced by an instance of an Indexed derived class.
I believe that the current implementation passes all unit tests for existing features when the base parser