fix: Adjust how WritableIndex.load creates new instances
The current implementation of WritableIndex.load assumed that
OffsetIndex and HierarchicalOffsetIndex took two parameters,
when in reality they took only one. This lead to an error that
signaled the caller automatically rebuilding the index every time,
never benefiting from the saved index.

This change corrects the way the container is created.